overlay_tool/requirements.txt
mss>=9.0.0
rapidocr>=3.3.0
onnxruntime>=1.16.0
pypinyin>=0.50.0
Pillow>=10.0.0
pyinstaller>=6.0.0
